Clean x86 & x64 Builds
Compiles without warnings on x86/x64, DebugAS/ReleaseAS

+++ source/MiniViewsLib/MainDialog/AboutTab.cpp
@@ -9,7 +9,7 @@ enum class Id : int
     ButtonManual
 };
 
-bool AboutWindow::CreateThis(HWND hParent, u32 windowId)
+bool AboutWindow::CreateThis(HWND hParent, u64 windowId)
 {
     if ( getHandle() != NULL )
         return SetParent(hParent);
