+++ source/MiniViewsLib/MainDialog/ViewsTab.cpp
@@ -19,7 +19,7 @@ enum class Id : int
 	CheckHideWhenSourceOnTop
 };
 
-bool ViewsWindow::CreateThis(HWND hParent, u32 windowId)
+bool ViewsWindow::CreateThis(HWND hParent, u64 windowId)
 {
 	if ( getHandle() != NULL )
 		return SetParent(hParent);
